fix(android): integer volumeLimitMl + shot snapshots from raw profile JSON (issue #23 sweep)
Follow-up to f9dbc1f6, from a sweep of the profile JSON boundary against
how the sibling apps type these values (reaprime: double + lenient
int/double/string parsing; Decenza: double + string-tolerant parse —
'de1app encodes numbers as strings'; de1app: untyped TCL). Every
ecosystem parser is lenient; our core now is too, and the canonical
wire semantic is whole ml.

1. volumeLimitMl is now an Int end-to-end on Android (model, SegmentEdit,
   both editors' steppers), so saves write `100` instead of `100.0` —
   matching the web (which already Math.round()s) and the core's u16.
   Decode stays tolerant of the floats older saves carry via
   LenientIntSerializer (JsonTransformingSerializer: float → rounded
   int), mirroring the core's lenient_uint; verified on-device by
   booting against a pre-switch customProfiles.json.

2. The sweep found a second, silent bug in the same class:
   captureCompletedShot re-encoded the THIN display model into
   cremaProfileToWire, and kotlinx omits fields sitting at their
   declared defaults — a segment target of 0, a dose of exactly 18 g
   (7g basket, Londinium Lever, Default Espresso, Damian's LRv2…) — so
   the strict core parse failed and the stored shot silently lost its
   embedded recipe (runCatching → null). Confirmed empirically: every
   dose-18 shot in the validation emulator's history has profile: null.
   The snapshot now converts from the FULL raw library JSON (the same
   source startShot uploads from) and logs on failure. Verified: a
   fresh replay shot on Damian's LRv2 (dose 18.0) embeds all 7 steps.

Also checked and clean: the community-v2 import path (already lenient
via Scalar — covers de1app string numbers, string version, nulls) and
the web (full-shape model, JSON.stringify has no default-omission).

assembleDebug green; 53 unit tests incl. new ProfileSegmentCodecTest.

+++ android/app/src/main/java/coffee/crema/profiles/CustomProfileStore.kt
@@ -240,7 +240,8 @@ data class SegmentEdit(
     val temp: Float?,
     /** Which sensor [temp] targets — `"coffee"` | `"water"`. */
     val tempSensor: String? = null,
-    val volume: Float? = null,
+    /** Per-segment volume limit, whole ml (wire `volumeLimitMl`, a core u16). */
+    val volume: Int? = null,
     /** Early-exit condition, or null = disabled. */
     val exit: SegmentExit? = null,
     /** Max limiter, or null = disabled. */
